# infra/bedrock/plan-management-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PlanManagement Action Group",
    "description": "Action group for managing learning plans in the Standards Database. Provides operations to create, update, retrieve, and delete personalized early learning plans.",
    "version": "1.0.0"
  },
  "paths": {
    "/createPlan": {
      "post": {
        "operationId": "createPlan",
        "summary": "Create a new learning plan",
        "description": "Creates a new personalized learning plan and persists it to the database. The plan is associated with the authenticated user and includes child profile information, interests, concerns, duration, and structured plan content.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "userId",
                  "childName",
                  "childAge",
                  "state",
                  "duration",
                  "content"
                ],
                "properties": {
                  "userId": {
                    "type": "string",
                    "description": "Authenticated user's Descope user ID"
                  },
                  "childName": {
                    "type": "string",
                    "description": "Child's first name"
                  },
                  "childAge": {
                    "type": "string",
                    "description": "Child's age or age range (e.g. 3, 4-5, infant)"
                  },
                  "state": {
                    "type": "string",
                    "description": "Two-letter state code (e.g. AZ, CA, TX)"
                  },
                  "interests": {
                    "type": "string",
                    "nullable": true,
                    "description": "Free-text description of the child's interests"
                  },
                  "concerns": {
                    "type": "string",
                    "nullable": true,
                    "description": "Free-text description of areas of concern"
                  },
                  "duration": {
                    "type": "string",
                    "description": "Plan duration (e.g. 1-week, 4-weeks, immediate)"
                  },
                  "content": {
                    "type": "string",
                    "description": "JSON string of the structured plan content (PlanContent object)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Plan created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "plan": {
                      "$ref": "#/components/schemas/PlanDetail"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing required fields or invalid content JSON",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/updatePlan": {
      "put": {
        "operationId": "updatePlan",
        "summary": "Update an existing learning plan",
        "description": "Updates the content of an existing learning plan. The plan must be owned by the specified user.",
        "parameters": [
          {
            "name": "planId",
            "in": "query",
            "required": true,
            "description": "UUID of the plan to update",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "description": "Authenticated user's Descope user ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "content"
                ],
                "properties": {
                  "planId": {
                    "type": "string",
                    "description": "UUID of the plan to update (alternative to query parameter)"
                  },
                  "userId": {
                    "type": "string",
                    "description": "Authenticated user's Descope user ID (alternative to query parameter)"
                  },
                  "content": {
                    "type": "string",
                    "description": "JSON string of the updated structured plan content (PlanContent object)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Plan updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "plan": {
                      "$ref": "#/components/schemas/PlanDetail"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing required fields or invalid content JSON",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Plan not found or not owned by user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/getPlan": {
      "get": {
        "operationId": "getPlan",
        "summary": "Retrieve a learning plan",
        "description": "Retrieves the full details of a learning plan by its ID. The plan must be owned by the specified user.",
        "parameters": [
          {
            "name": "planId",
            "in": "query",
            "required": true,
            "description": "UUID of the plan to retrieve",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "description": "Authenticated user's Descope user ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plan retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "plan": {
                      "$ref": "#/components/schemas/PlanDetail"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing required parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Plan not found or not owned by user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/deletePlan": {
      "delete": {
        "operationId": "deletePlan",
        "summary": "Delete a learning plan",
        "description": "Deletes a learning plan by its ID. The plan must be owned by the specified user.",
        "parameters": [
          {
            "name": "planId",
            "in": "query",
            "required": true,
            "description": "UUID of the plan to delete",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "description": "Authenticated user's Descope user ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plan deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "description": "Whether the deletion was successful"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing required parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Plan not found or not owned by user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "PlanDetail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "UUID of the plan"
          },
          "userId": {
            "type": "string",
            "description": "Descope user ID of the plan owner"
          },
          "childName": {
            "type": "string",
            "description": "Child's first name"
          },
          "childAge": {
            "type": "string",
            "description": "Child's age or age range"
          },
          "state": {
            "type": "string",
            "description": "Two-letter state code"
          },
          "interests": {
            "type": "string",
            "nullable": true,
            "description": "Child's interests"
          },
          "concerns": {
            "type": "string",
            "nullable": true,
            "description": "Areas of concern"
          },
          "duration": {
            "type": "string",
            "description": "Plan duration"
          },
          "content": {
            "type": "object",
            "description": "Structured plan content",
            "properties": {
              "sections": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/PlanSection"
                }
              },
              "summary": {
                "type": "string",
                "description": "Agent-generated summary of the plan"
              }
            }
          },
          "status": {
            "type": "string",
            "description": "Plan status (active, archived)"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the plan was created"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the plan was last updated"
          }
        }
      },
      "PlanSection": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "description": "Section label (e.g. Week 1, Immediate Suggestions)"
          },
          "description": {
            "type": "string",
            "description": "Optional section introduction"
          },
          "activities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlanActivity"
            }
          }
        }
      },
      "PlanActivity": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Activity title"
          },
          "description": {
            "type": "string",
            "description": "Detailed activity description"
          },
          "indicatorCode": {
            "type": "string",
            "description": "Standard indicator code (e.g. MA.PK.1.2)"
          },
          "indicatorDescription": {
            "type": "string",
            "description": "Full text of the standard indicator"
          },
          "domain": {
            "type": "string",
            "description": "Learning domain (e.g. Mathematics)"
          },
          "strand": {
            "type": "string",
            "description": "Strand within the domain (e.g. Number Sense)"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message describing what went wrong"
          }
        }
      }
    }
  }
}